FEATURE - friend list to be used for checking people in and inviting them to use the app. friends become User entities even if they are not registered

+++ wineguide/controllers/TutorialDetailsVC.h
@@ -10,7 +10,7 @@
 
 @interface TutorialDetailsVC : UIViewController
 
-@property (assign, nonatomic) int index;
+@property (assign, nonatomic) NSInteger index;
 @property (strong, nonatomic) IBOutlet UILabel *screenInstructionText;
 
 @end
